+++ src/sdc11073/roles/protocols.py
@@ -6,7 +6,7 @@ if TYPE_CHECKING:
 
     from sdc11073.mdib import ProviderMdib
     from sdc11073.mdib.descriptorcontainers import AbstractOperationDescriptorProtocol
-    from sdc11073.mdib.transactions import RtDataMdibUpdateTransaction
+    from sdc11073.mdib.transactionsprotocol import RtDataMdibUpdateTransaction
     from sdc11073.provider.operations import OperationDefinitionBase
     from sdc11073.provider.sco import AbstractScoOperationsRegistry
     from sdc11073.xml_types.pm_types import ComponentActivation
